refactor: split trip planning slice state and cache helpers
